reafactor: 계산 결과를 int -> double로 다시 수정하고 0으로 나눌 수 없는 예외 추가

+++ src/main/java/study/Application.java
@@ -13,7 +13,7 @@ public class Application {
 		try {
 			String expression = InputView.inputExpression();
 			String[] tokens = expression.split(" ");
-			int result = Calculator.calculate(tokens);
+			double result = Calculator.calculate(tokens);
 			OutputView.printResult(result);
 		} catch (Exception e) {
 			System.out.println("잘못된 식입니다.");
