+++ src/main/java/view/OutputView.java
@@ -1,7 +1,7 @@
 package view;
 
 public class OutputView {
-	public static void printResult(int result) {
+	public static void printResult(double result) {
 		System.out.println("결과 값: " + result);
 	}
 }
